ParsingUtils update
