+++ src/main/java/org/jetbrains/bsp/bazel/server/BUILD
@@ -8,14 +8,12 @@ java_library(
         "//src/main/java/org/jetbrains/bsp/bazel/commons",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bazel",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/data",
-        "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/utils",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bep",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bsp",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bsp/impl",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bsp/resolvers",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bsp/services",
         "//src/main/java/org/jetbrains/bsp/bazel/server/logger",
-        "//src/main/java/org/jetbrains/bsp/bazel/server/utils",
         "@io_bazel//src/main/protobuf:build_java_proto",
         "@io_bazel//third_party/grpc:grpc-jar",
         "@maven//:ch_epfl_scala_bsp4j",
