+++ src/main/java/org/jetbrains/bsp/bazel/server/bsp/BUILD
@@ -11,7 +11,7 @@ java_library(
         "//src/main/java/org/jetbrains/bsp/bazel/server/bazel/params",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bep",
         "//src/main/java/org/jetbrains/bsp/bazel/server/bsp/resolvers",
-        "//src/main/java/org/jetbrains/bsp/bazel/server/utils",
+        "//src/main/java/org/jetbrains/bsp/bazel/server/bsp/utils",
         "@io_bazel//src/main/protobuf:build_java_proto",
         "@io_bazel//third_party/grpc:grpc-jar",
         "@maven//:ch_epfl_scala_bsp4j",
